shrimech

+++ ft_printf.c
@@ -6,7 +6,7 @@
 /*   By: shrimech <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/11/15 11:09:34 by shrimech          #+#    #+#             */
-/*   Updated: 2024/11/26 18:47:11 by shrimech         ###   ########.fr       */
+/*   Updated: 2024/11/26 18:49:05 by shrimech         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -23,7 +23,7 @@ int	ft_putnbr_hex(unsigned long n)
 	return (i);
 }
 
-int	ft_putnbr_hex_upper(unsigned int n)
+int	ft_putnbr_hex_upper(unsigned long n)
 {
 	int	i;
